Refactor code in clCreateProgramWithBuiltinKernels()
Reuse code which is also used by other compiler-related calls
to create a program with binaries.

+++ lib/CL/pocl_build.c
@@ -814,7 +814,7 @@ compile_and_link_program(int compile_program,
               program->kernel_meta = calloc (program->num_kernels,
                                              sizeof (pocl_kernel_metadata_t));
 
-              for (int i = 0; i < program->num_kernels; ++i)
+              for (size_t i = 0; i < program->num_kernels; ++i)
                 {
                   device->ops->get_builtin_kernel_metadata (
                       device->data, program->builtin_kernel_names[i],
